chore: move more database stuff
